DDF-3551 Added a new username parameter to the QueryDeliveryService API
Services that may need to access additional data about a delivery method cannot just use the delivery ID to identify the additional data; only a username and delivery ID together are guaranteed to uniquely identify a delivery method. Therefore, the username was added to be used in conjunction with the delivery ID to access additional information from other sources outside of user preferences.