DDF-3551 Refactored to match intended parameter and response format
The frontend and backend were expecting different data formats for delivery parameters; the frontend was expecting a "fields" key to correspond to a list of objects with "name" and "type" while the backend expected a "deliveryType" key to correspond to one object where names were keys and types were values. In a previous commit, the backnd was changed to fit the frontend on a time crunch; now proper refactoring has made changed the frontend to fot the backend instead.

DDF-3551 [WIP] QueryCourier revisions

This commit is a work in progress; it does not compile!

The query delivery API is being revised based on suggestions in an attempt to make it stable and reusable enough to include in the Catalog core API.

DDF-3551 [WIP] Backend cleanup and logging fix

This commit fixes a bug that broke the logging config and cleans up some backend code.

This commit will compile successfully, but does not work properly! The frontend needs some changes to correspond to some changes in the backend.

DDF-3551 [WIP] Filter out Ignite stdout printing

This commit implements one strategy for preventing Ignite from spitting logs into the Karaf console (stdout) by setting System.out to a new PrintStream that checks the stack trace and logs message from Ignite to a Logger or simply prints the message otherwise.

Issues with the previous commit still stand.

DDF-3551 Resolved cherry pick issues

A couple issues that should have been resolved during the cherry-pick process ended up in the previous commit. Fixed!

DDF-3551 Refactor completion and bug fixes!

This commit finalizes the refactoring of the deliveryParameters format (formerly `fields`) and fixes bugs related to scheduling (mainly caused by the incomplete refactor in previous commits).